YD-497 Processed review comments
* Moved DeviceRegistrationRequestDto and DeviceUpdateRequestDto to their own files
* Simplified DeviceManagementTest.'Richard adds new device; Bob gets the update'
* Introduced updateResourceWithPassword and used it
* Used deleteResourceWithPassword instead of just deleteResource
* Swapped method names in DeviceService: deleteDevice --> deleteDeviceWithoutBuddyNotification and deleteDeviceAndNotifyBuddies --> deleteDevice. The exception now has a special name
* Improved test name in DeviceServiceTest
