+++ appservice/src/main/java/nu/yona/server/subscriptions/rest/UserController.java
@@ -58,9 +58,9 @@ import nu.yona.server.crypto.seckey.CryptoSession;
 import nu.yona.server.crypto.seckey.SecretKeyUtil;
 import nu.yona.server.device.rest.DeviceController;
 import nu.yona.server.device.service.DeviceBaseDto;
+import nu.yona.server.device.service.DeviceRegistrationRequestDto;
 import nu.yona.server.device.service.DeviceService;
 import nu.yona.server.device.service.UserDeviceDto;
-import nu.yona.server.device.service.UserDeviceDto.DeviceRegistrationRequestDto;
 import nu.yona.server.exceptions.ConfirmationException;
 import nu.yona.server.exceptions.InvalidDataException;
 import nu.yona.server.exceptions.YonaException;
